re-organizes packaging
