+++ src/test/java/networking/stub/StubNetworkTest.java
@@ -1,4 +1,4 @@
-package networking;
+package networking.stub;
 
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
@@ -7,6 +7,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import model.Entity;
 import model.exceptions.LightChainNetworkingException;
 import network.Conduit;
+import networking.MockEngine;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import unittest.fixtures.EntityFixture;
